Changed up the solution to be a stub framework more suitable for milestone 1
I took the example framework and stripped out the windowing stuff to
leave it with just stubs, which is more suitable for the bare-bones
framework we need to hand in for milestone 1. I also implemented the
timing system I'd talked about, as that's another requirement Dr.
Stephenson mentioned. I brought in SDL in the process, since
SDL_GetTicks is what we'll use so might as well use it here.