Renderables set up in a JSON file that gets read
Renderables are now in a JSON file instead of hard-coded, and matched to
a name in a map. This is the first step in getting the tile files done.
I'm trying to add a new model but having trouble with exporting it from
Blender in a way the parser likes.